requirements.txt
gunicorn
Flask  
bs4
requests
flask-sqlalchemy
flask-restful
python-dotenv
flask-migrate
flask-marshmallow
marshmallow-sqlalchemy
passlib
flask-jwt-extended
psycopg2 